Add abstract class for MustFightBattle & NonFightingBattle to consolidate duplicated code
